requirements.txt
discord.py
openai
python-dotenv
# spacy
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz
